Refactor save file
