+++ src/CodeSharing.Server/Repositories/Intefaces/IUploadRepository.cs
@@ -1,0 +1,6 @@
+namespace CodeSharing.Server.Repositories.Intefaces;
+
+public interface IUploadRepository
+{
+    Task<string> SaveFile(IFormFile file);
+}
